requirements.txt
python=3.7.6
pip=20.0.2
matplotlib=3.1.3
pytorch==1.4.0
pip:
numpy==1.18.1
opencv-python==4.2.0.32
torchvision==0.5